Extracting invites' email sending into an InviteEmail operation object
Following the service/operation object pattern (for instance, see
http://trailblazer.to/gems/operation/2.0/), we should be able to send e-mail
invites decoupled from a particular controller action.

This also allows one to use a different mailer class from the original. Useful
when you want to configure the e-mail for a particular invitable.
