+++ app/controllers/invitation/invites_controller.rb
@@ -13,7 +13,7 @@ module Invitation
   # * override after_invite_existing_user or after_invite_new_user
   #
   class InvitesController < ApplicationController
-    
+
     def new
       @invite = InviteForm.new(invite_params)
       render template: 'invites/new'
@@ -24,12 +24,8 @@ module Invitation
     #   invite: { invitable_id, invitable_type, email or emails:[] }
     #
     def create
-      failures = []
       invites = InviteForm.new(invite_params).build_invites(current_user)
-      ActiveRecord::Base.transaction do
-        invites.each { |invite| invite.save ? do_invite(invite) : failures << invite.email }
-      end
-
+      failures = Invitation::InviteEmails.new(invites).send_invites
       logger.info "!!!!!!!!!!!!!!!!!!!!! INSIDE CREATE: current_user: #{current_user.inspect}"
       respond_to do |format|
         format.html do
@@ -55,21 +51,6 @@ module Invitation
       end
     end
 
-    private
-
-    # Override this if you want to do something more complicated for existing users.
-    # For example, if you have a more complex permissions scheme than just a simple
-    # has_many relationship, enable it here.
-    def after_invite_existing_user(invite)
-      # Add the user to the invitable resource/organization
-      invite.invitable.add_invited_user(invite.recipient)
-    end
-
-    # Override if you want to do something more complicated for new users.
-    # By default we don't do anything extra.
-    def after_invite_new_user(invite)
-    end
-
     # After an invite is created, redirect the user here.
     # Default implementation doesn't return a url, just the invitable.
     def url_after_invite(invite)
@@ -78,29 +59,6 @@ module Invitation
 
     def invite_params
       params[:invite] ? params.require(:invite).permit(:invitable_id, :invitable_type, :email, emails: []) : {}
-    end
-
-    # Invite user by sending email.
-    # Existing users are granted permissions via #after_invite_existing_user.
-    # New users are granted permissions via #after_invite_new_user, currently a null op.
-    def do_invite(invite)
-      if invite.existing_user?
-        deliver_email(InviteMailer.existing_user(invite))
-        after_invite_existing_user(invite)
-        invite.save
-      else
-        deliver_email(InviteMailer.new_user(invite))
-        after_invite_new_user(invite)
-      end
-    end
-
-    # Use deliver_later from rails 4.2+ if available.
-    def deliver_email(mail)
-      if mail.respond_to?(:deliver_later)
-        mail.deliver_later
-      else
-        mail.deliver
-      end
     end
   end
 end
